exam/2023_Q2/andruc/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
idna==3.4
nose==1.3.7
numpy==1.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
PySide6==6.4.0.1
PySide6-Addons==6.4.0.1
PySide6-Essentials==6.4.0.1
PyYAML==6.0
Random-Word==1.0.7
requests==2.28.1
shiboken6==6.4.0.1
urllib3==1.26.12
validators==0.20.0
